Add bigint support for RealTime indices

+++ lib/thinking_sphinx/real_time/index/template.rb
@@ -8,7 +8,7 @@ class ThinkingSphinx::RealTime::Index::Template
   def apply
     add_field class_column, :sphinx_internal_class_name
 
-    add_attribute :id,          :sphinx_internal_id,    :integer
+    add_attribute :id,          :sphinx_internal_id,    :bigint
     add_attribute class_column, :sphinx_internal_class, :string, :facet => true
     add_attribute 0,            :sphinx_deleted,        :integer
   end
